added eth to canto address function, query vote by proposal id and voter
